third_party/WebKit/LayoutTests/accessibility/aria-disabled-expected.txt

 
link  submit
This tests that the aria-disabled attribute works. The text field should not be enabled.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS textField.isEnabled is false
PASS textField.isEnabled is true
PASS group.isEnabled is false
PASS link.isEnabled is false
PASS button.isEnabled is true
PASS successfullyParsed is true

TEST COMPLETE
